% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Catalog.R
\name{mod_Catalog_server}
\alias{mod_Catalog_server}
\title{Catalog Server Functions}
\usage{
mod_Catalog_server(id)
}
\arguments{
\item{id}{A unique identifier for the shiny module.}
}
\description{
This function defines server logic for the Catalog module in a Shiny application.
It handles file input and data processing for catalog entries.
}
